Move Lor Match DTOs to Lor namespace
