+++ tests/Unit/DTO/Lor/MatchDTOTest.php
@@ -2,12 +2,12 @@
 
 declare(strict_types=1);
 
-namespace Riot\Tests\Unit\DTO;
+namespace Riot\Tests\Unit\DTO\Lor;
 
 use PHPUnit\Framework\TestCase;
-use Riot\DTO\InfoDTO;
-use Riot\DTO\MatchDTO;
-use Riot\DTO\MetadataDTO;
+use Riot\DTO\Lor\InfoDTO;
+use Riot\DTO\Lor\MatchDTO;
+use Riot\DTO\Lor\MetadataDTO;
 
 final class MatchDTOTest extends TestCase
 {
